requirements.txt.txt
streamlit
pandas
streamlit-gsheets-connection